requirements.txt
streamlit
plotly.express

